+++ projects/ngx-qgrid-plugins/src/lib/pagination/pager.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input, OnInit, OnChanges, ChangeDetectionStrategy, ChangeDetectorRef } from '@angular/core';
 import { GridPlugin, StateAccessor } from '@qgrid/ngx';
 import { PagerPlugin } from '@qgrid/plugins/pager/pager.plugin';
-import { PaginationState } from '@qgrid/core/pagination/pagination.state';
+import { PaginationState, PaginationStateMode } from '@qgrid/core/pagination/pagination.state';
 
 @Component({
 	selector: 'q-grid-pager',
@@ -12,11 +12,11 @@ import { PaginationState } from '@qgrid/core/pagination/pagination.state';
 export class PagerComponent implements OnInit, OnChanges {
 	private pagerState = this.stateAccessor.setter(PaginationState);
 
-	context: { $implicit: PagerPlugin, plugin: PagerComponent };
+	context: { $implicit: PagerPlugin };
 
-	@Input('size') set paginationSize(size: number) { this.pagerState({ size }); }
-	@Input('sizeList') set paginationSizeList(sizeList: number[]) { this.pagerState({ sizeList }); }
-	@Input('showCount') showCount = false;
+	@Input('size') set size(size: number) { this.pagerState({ size }); }
+	@Input('sizeList') set sizeList(sizeList: number[]) { this.pagerState({ sizeList }); }
+	@Input('mode') set mode(mode: PaginationStateMode) { this.pagerState({ mode }); }
 
 	constructor(
 		private plugin: GridPlugin,
@@ -32,8 +32,9 @@ export class PagerComponent implements OnInit, OnChanges {
 
 	ngOnInit() {
 		const { model, observe } = this.plugin;
+
 		const pager = new PagerPlugin(this.plugin);
-		this.context = { $implicit: pager, plugin: this };
+		this.context = { $implicit: pager };
 
 		observe(model.paginationChanged)
 			.subscribe(() => this.cd.detectChanges());