+++ projects/qgrid-core/pagination/pagination.state.d.ts
@@ -1,3 +1,5 @@
+export declare type PaginationStateMode = 'showPages' | 'showRows';
+
 /**
  * A class that allows to split data to pages, also virtual scroll use some options from here.
  */
@@ -21,6 +23,8 @@ export declare class PaginationState {
 	 * Count of total rows.
 	 */
 	count: number;
+
+	mode: PaginationStateMode;
 
 	/**
 	 * List of `model name: [model properties]` pairs to reset pagination current property to 0.